Initial Resources Cut
This commit is the first version of json resources that will drive
the end-product.  Doing these in JSON.

There are a few TODOs in there that denote some design concerns. That
said, I'm just kinda going, and will likely write a something that
compiles all of the disparate jsons together.